Relocate useLinkEditModal to @editor
